+++ apps/vaults-v3/components/SettingsPopover.tsx
@@ -37,22 +37,22 @@ function MaxLossSection(): ReactElement {
 			</legend>
 			<div className={'flex flex-row space-x-2 pt-2'}>
 				<button
-					onClick={(): void => set_maxLoss(100n)}
+					onClick={(): void => set_maxLoss(5n)}
 					className={`flex h-10 items-center rounded-lg border bg-neutral-100 px-1.5 py-2 ${
-						maxLoss === 100n ? 'border-neutral-900/40' : 'border-transparent'
+						maxLoss === 5n ? 'border-neutral-900/40' : 'border-transparent'
 					}`}>
-					<p className={'font-number px-2 text-center text-neutral-900 '}>{'1%'}</p>
+					<p className={'font-number px-2 text-center text-neutral-900 '}>{'0.05%'}</p>
 				</button>
 				<button
-					onClick={(): void => set_maxLoss(200n)}
+					onClick={(): void => set_maxLoss(10n)}
 					className={`flex h-10 items-center rounded-lg border bg-neutral-100 px-1.5 py-2 ${
-						maxLoss === 200n ? 'border-neutral-900/40' : 'border-transparent'
+						maxLoss === 10n ? 'border-neutral-900/40' : 'border-transparent'
 					}`}>
-					<p className={'font-number px-2 text-center text-neutral-900 '}>{'2%'}</p>
+					<p className={'font-number px-2 text-center text-neutral-900 '}>{'0.1%'}</p>
 				</button>
 				<div
-					className={`flex h-10 w-full min-w-[72px] items-center rounded-lg border bg-neutral-100 px-0 py-4 md:min-w-[160px] ${
-						maxLoss !== 100n && maxLoss !== 200n ? 'border-neutral-900/40' : 'border-transparent'
+					className={`flex h-10 w-full min-w-[72px] items-center rounded-lg border bg-neutral-100 px-0 py-4 ${
+						maxLoss !== 5n && maxLoss !== 10n ? 'border-neutral-900/40' : 'border-transparent'
 					}`}>
 					<input
 						type={'number'}
@@ -74,6 +74,13 @@ function MaxLossSection(): ReactElement {
 					<p className={'font-number mt-1 pr-2 text-neutral-900/60'}>{'%'}</p>
 				</div>
 			</div>
+			{maxLoss >= 100n ? (
+				<div className={'pt-2'}>
+					<p className={'rounded-lg border-red-900 bg-red-200 p-2 text-xs text-red-900'}>
+						{'⚠ You have selected a very high max loss. Please make sure this is correct.'}
+					</p>
+				</div>
+			) : null}
 		</div>
 	);
 }